Added in enemy model, Recreated the main menu with a moving scene
Ive been working on an enemy model.

i recreated the main menu with a moving camera in a scene with the main character on stage. i did this as the flat image on the canvas looked a little pixilated and tought it needed changing

I added a scrpit for the camera with varibles we can adjust